fix: page and size properties all set to long type

+++ src/TrendyolSharp/Models/Finance/Response/ResponseGetOtherFinancials.cs
@@ -4,8 +4,8 @@ namespace TrendyolSharp.Models.Finance.Response
 {
   public sealed class ResponseGetOtherFinancials
   {
-    public int Page { get; set; }
-    public int Size { get; set; }
+    public long Page { get; set; }
+    public long Size { get; set; }
     public long TotalPages { get; set; }
     public long TotalElements { get; set; }
     public List<OtherFinancialsTransactionData> Content { get; set; }
