+++ src/TrendyolSharp/Models/Marketplace/Response/ResponseGetClaims.cs
@@ -6,8 +6,8 @@ namespace TrendyolSharp.Models.Marketplace.Response
   {
     public long TotalElements { get; set; }
     public long TotalPages { get; set; }
-    public int Page { get; set; }
-    public int Size { get; set; }
+    public long Page { get; set; }
+    public long Size { get; set; }
     public List<Claim> Content { get; set; }
   }
 }
